// ML_LandEvaluation_Application/src/data/models/state.min-max.model.ts
export class StateMinMax {
    PopulationMax: number = 7171646;
    RDPIMax: number = 39955;
    EmploymentMax: number = 3859137;
    PopulationMin: number = 6280362;
    RDPIMin: number = 30844;
    EmploymentMin: number = 3181571;
    PriceMax: number = 247449.5833;
    PriceMin: number = 140399.6667;
}