// ML_LandEvaluation_Application/src/data/models/county.min-max.model.ts
export class CountyMinMax {
    PriceMax: number;
    PriceMin: number;
    CrimeIncidentsMax: number;
    CrimeIncidentsMin: number;
    PopulationMax: number;
    PopulationMin: number;
    PerCapitaPersonalIncomeMax: number;
    PerCapitaPersonalIncomeMin: number;
    MedianHouseholdIncomeMax: number;
    MedianHouseholdIncomeMin: number;
}